requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
python-dotenv
pydantic
typing-extensions
sentence-transformers
faiss-cpu
google-generativeai
PyPDF2
python-docx
python-multipart
scikit-learn
